requirements.txt
python-telegram-bot==20.0
mysql-connector-python==8.0.30
python-dotenv==1.0.0
